Android project to control a Bluetooth robot over internet.
It will allow a user to command a robot using an Android device.
Communication between device and robot is via Bluetooth and the device
uses internet to receive commands from user. Also the device is making
use of the camera to send images so that the user can see the robot
path.